feat(frontend): delete duplicate files and rename
 - Delete all duplicate files of controls/ infraction
 - rename 2 without not 2. modify imports
