+++ backend/src/test/kotlin/fr/gouv/gmampa/rapportnav/mocks/mission/action/MissionNavActionEntityMock.kt
@@ -7,6 +7,7 @@ import fr.gouv.dgampa.rapportnav.domain.entities.mission.nav.control.ControlMeth
 import fr.gouv.dgampa.rapportnav.domain.entities.mission.nav.status.ActionStatusReason
 import fr.gouv.dgampa.rapportnav.domain.entities.mission.nav.status.ActionStatusType
 import fr.gouv.dgampa.rapportnav.domain.entities.mission.v2.MissionNavActionEntity
+import fr.gouv.dgampa.rapportnav.domain.entities.mission.v2.TargetEntity2
 import java.time.Instant
 import java.util.*
 
@@ -47,6 +48,7 @@ object MissionNavActionEntityMock {
         nbAssistedVesselsReturningToShore: Int? = null,
         status: ActionStatusType? = null,
         reason: ActionStatusReason? = null,
+        targets: List<TargetEntity2>? = null
     ): MissionNavActionEntity {
         return MissionNavActionEntity(
             missionId = missionId?: 761,
@@ -83,7 +85,8 @@ object MissionNavActionEntityMock {
             nbOfVesselsTrackedWithoutIntervention = nbOfVesselsTrackedWithoutIntervention,
             nbAssistedVesselsReturningToShore = nbAssistedVesselsReturningToShore,
             reason = reason,
-            status = status
+            status = status,
+            targets = targets
         )
     }
 }
