+++ frontend/src/v2/features/common/types/mission-action-data.ts
@@ -1,9 +1,3 @@
-import {
-  ControlAdministrative,
-  ControlGensDeMer,
-  ControlNavigation,
-  ControlSecurity
-} from '@common/types/control-types'
 import { Target } from './target-types'
 
 export interface MissionActionData {
@@ -12,9 +6,5 @@ export interface MissionActionData {
   observations?: string
   longitude?: number
   latitude?: number
-  controlSecurity?: ControlSecurity
-  controlGensDeMer?: ControlGensDeMer
-  controlNavigation?: ControlNavigation
-  controlAdministrative?: ControlAdministrative
   targets?: Target[]
 }
