+++ frontend/src/v2/features/mission-action/components/elements/mission-action-item-control.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react'
 import { MissionAction } from '../../../common/types/mission-action'
 import { MissionSourceEnum } from '../../../common/types/mission-types'
-import MissionActionItemEnvControl2 from './mission-action-item-env-control2'
+import MissionActionItemEnvControl from './mission-action-item-env-control'
 import MissionActionItemFishControl from './mission-action-item-fish-control'
 import MissionActionItemNavControl from './mission-action-item-nav-control'
 
@@ -12,7 +12,7 @@ const MissionActionItemControl: FC<{
 }> = ({ action, onChange, isMissionFinished }) => {
   switch (action.source) {
     case MissionSourceEnum.MONITORENV:
-      return <MissionActionItemEnvControl2 action={action} onChange={onChange} />
+      return <MissionActionItemEnvControl action={action} onChange={onChange} />
     case MissionSourceEnum.MONITORFISH:
       return <MissionActionItemFishControl action={action} onChange={onChange} isMissionFinished={isMissionFinished} />
     default:
