+++ frontend/src/v2/features/mission-control/components/elements/mission-control-form.tsx
@@ -4,8 +4,8 @@ import { FieldArray, FieldArrayRenderProps, FieldProps, Formik } from 'formik'
 import { FC } from 'react'
 import { Panel, Stack } from 'rsuite'
 import { Control } from '../../../common/types/target-types'
-import MissionInfractionList2 from '../../../mission-infraction/components/elements/mission-infraction-list-form2'
-import { ControlModelInput, useControl } from '../../hooks/use-control'
+import MissionInfractionList from '../../../mission-infraction/components/elements/mission-infraction-list'
+import { ControlInput, useControl } from '../../hooks/use-control'
 import MissionActionControlFormikMultiRadio from '../ui/mission-control-fomik-multi-radio'
 import { MissionControlTitle } from '../ui/mission-control-title'
 import MissionControlUnitConfirm from '../ui/mission-control-unit-confirm'
@@ -30,7 +30,7 @@ const MissionControlForm: FC<MissionControlModelFormProps> = ({ name, controlTyp
         <Formik initialValues={initValue} onSubmit={handleSubmit} enableReinitialize>
           {formik => (
             <>
-              <FormikEffect onChange={nextValues => handleSubmit(nextValues as ControlModelInput)} />
+              <FormikEffect onChange={nextValues => handleSubmit(nextValues as ControlInput)} />
               <Stack direction="column" alignItems="flex-start" spacing="1rem" style={{ width: '100%' }}>
                 <Stack.Item style={{ width: '100%' }}>
                   <MissionControlUnitConfirm
@@ -61,11 +61,7 @@ const MissionControlForm: FC<MissionControlModelFormProps> = ({ name, controlTyp
                     <Stack.Item style={{ width: '100%' }}>
                       <FieldArray name="infractions">
                         {(fieldArray: FieldArrayRenderProps) => (
-                          <MissionInfractionList2
-                            name="infractions"
-                            fieldArray={fieldArray}
-                            controlType={controlType}
-                          />
+                          <MissionInfractionList name="infractions" fieldArray={fieldArray} controlType={controlType} />
                         )}
                       </FieldArray>
                     </Stack.Item>
