+++ frontend/src/v2/features/mission-control/components/ui/mission-control-fish-other-infractions-section.tsx
@@ -2,7 +2,7 @@ import { Label } from '@mtes-mct/monitor-ui'
 import React from 'react'
 import { Stack } from 'rsuite'
 import { MissionFishActionData } from '../../../common/types/mission-action'
-import MissionInfractionSummary from '../../../mission-infraction/components/ui/mission-infraction-summary'
+import MissionInfractionFishSummary from '../../../mission-infraction/components/elements/mission-infraction-fish-summary'
 
 interface MissionControlFishOtherSectionProps {
   action: MissionFishActionData
@@ -15,7 +15,7 @@ const MissionControlFishOtherInfractionsSection: React.FC<MissionControlFishOthe
         <Label>Autres infractions</Label>
       </Stack.Item>
       <Stack.Item style={{ width: '100%' }}>
-        <MissionInfractionSummary
+        <MissionInfractionFishSummary
           title="Autres infractions"
           isActionDisabled={true}
           infractions={action.otherInfractions ?? []}
