+++ frontend/src/v2/features/mission-infraction/components/elements/mission-infraction-list.tsx
@@ -1,8 +1,8 @@
 import { ControlType } from '@common/types/control-types'
-import { Infraction } from '@common/types/infraction-types'
 import { FieldArrayRenderProps } from 'formik'
-import MissionInfractionFormNew from './mission-infraction-form-new'
-import MissionInfractionItem from './mission-infraction-item'
+import { Infraction } from '../../../common/types/target-types'
+import MissionInfractionFormNew2 from './mission-infraction-form-new'
+import MissionInfractionItem2 from './mission-infraction-item'
 
 const DEFAULT_MAX = 1
 export interface MissionInfractionListProps {
@@ -24,7 +24,7 @@ const MissionInfractionList: React.FC<MissionInfractionListProps> = ({ max, name
   return (
     <div>
       {fieldArray.form.values.infractions?.map((infraction: Infraction, index: number) => (
-        <MissionInfractionItem
+        <MissionInfractionItem2
           name={name}
           index={index}
           key={`${name}-index`}
@@ -32,7 +32,7 @@ const MissionInfractionList: React.FC<MissionInfractionListProps> = ({ max, name
           handleRemove={handleRemove}
         />
       ))}
-      <MissionInfractionFormNew
+      <MissionInfractionFormNew2
         onSubmit={handleAdd}
         controlType={controlType}
         isDisabled={fieldArray.form.values?.infractions?.length >= (max ?? DEFAULT_MAX)}
