+++ frontend/src/v2/features/mission-infraction/types/infraction-input.tsx
@@ -1,5 +1,4 @@
 import { GearInfraction, LogbookInfraction, OtherInfraction, SpeciesInfraction } from '@common/types/fish-mission-types'
-import { Infraction } from '@common/types/infraction-types'
 
-export type FishNavInfraction = LogbookInfraction | GearInfraction | SpeciesInfraction | OtherInfraction | Infraction
+export type FishInfraction = LogbookInfraction | GearInfraction | SpeciesInfraction | OtherInfraction
 //TODO: update Fish infraction to add id, natinf to natinfs plural in array, transfert comment into observations
